Tweaked TorusPredPreyTask
Completed TorusEvolvedPredatorVsStaticPreyTask so that predators are
evolved (as many as specified in parameters) and the prey are static.
The user specifies which of the agents (the prey or the predators) get
evolved by specifying which task will be used. I then created a
super/parent class called "TorusPredPreyTask" which contained most of
the code in the two specific children classes (extended by the two
specific evolving prey or predator classes) so that I was then able to
delete almost all of the code in the children, leaving only a
specification to the parent class of which type of agent is being
evolved and methods providing a list of the prey and predator
controllers (one is static and the other is evolved according to
parameter specifications).