Created the TorusPredPrey controller PreyFleeClosestPredatorController
Created PreyFleeClosestPredatorController, but it still is not being
used anywhere. I still need to create a command line parameter that
allows a user to switch between using the "PreyFleeAllPredators"
controller and the new "PreyFleeClosestPredator" controller as the
static prey agents when predators are evolved instead of just always
using the PreyFleeAllPredators controller by default.