Created
Created the SpreadOutAggressive predator controller and implemented the
parameter so that the user can switch between various static predators
while evolving prey. The SpreadOutAggressive controller still needs a
lot of work as far as the efficiency of the predators though, and I will
work on weighting the different functions of the predator (spreading out
from one another versus getting close to prey) better.