Fixed image match task
Now performs closer to CPPN expectations. Removed a few functions that
produced poor results. Added a scaling function that moved center of
image to actual center. Also scaled image to a -1 to 1. Also added a
main method for testing purposes. Added a distance from center and a
bias of 1 parameter to CPPN inputs. Also changed outputs to be decoded
as HSB values, not RGB. Finally, added a helper save function to allow
clients to save a bmp of the images they create. Also this commit
includes some of the cooler pictures created with the CPPN.