began new fitness function for predators and multiple monitor inputs for predators
began new fitness function for predators and multiple monitor inputs for
predators (multiple inputs for prey still needs to be added). Created
another predator fitness function, called
PredatorLowAverageDistanceFromPreyObjective, in attempt to ultimately
make a stronger one that is strong enough to capture the prey when they
are using the flee closest predator static prey controller. This fitness
function will reward the lowest average distances of predators to preys.
This should help me to implement the strong fitness functions for the
predators and the prey that were described in the A. Rawal and P.
Rajagopalan paper.