"Fixed" contract violation by comparator
I've made the contract violation error from sorting go away by using
Java's legacy sorting method. I'm not sure how my ObjectiveComparator is
violating the contract, but according to TimSort, it is. The better
solution would be to fix the comparator, but this change suppresses the
error and makes evolution behave the same way it did before ... despite
the supposed contract violation, the code seems to work.