Torus predPrey monitorInput fix, made another Pred fitness function, generalized dimensions
fixed a bug with the monitorInputs disposing monitors incorrectly after
each evaluation. Created a new predator fitness function
"PredatorMinimizeDistanceFromPreyObjective" which encourages that the
predators end up as close to the prey as possible by the end of the
game. Created parameters for the grid world dimensions so that it is not
hard coded as 100 by 100