CPPN problem fixed!
The problem was the absolute value function! This is the only activation
function whose value is not restricted to a finite range. This was
causing network activations to rise to the point where they reached
infinity, and once you introduce that you get all kinds of problems like
NaN values. These NaN values made images look terrible, ruined the image
match fitness function, but also broke the "contract" of sorting
comparators, causing the exceptions with TimSort. I was able to remove
the system property to use the legacy sorting method. I found the error
by adding tons of asserts to the code. Since asserts only run when a
command line parameter is set, I'm going to go ahead and leave them in
... they were really helpful!

With this fix, the French Flag can be evolved again. The absolute value
function must have been disabled the time I initially evolved it, but I
suppose it got reactivated in a commit that someone else made.

Note: random CPPNs still produce boring black images sometimes, but this
is reasonable.