Finished up the Torus PredPrey RRM paper fitness functions
Finished up the Torus PredPrey RRM paper fitness functions by creating a
static method in my predatorMinimizeDistanceFromPreyObjective which
found distance from each predator to each prey and summed those up then
found the sum of those sums (which was already being done, but was moved
to a static method) and then I was able to call this method from the RRM
predator method to implement that. Because of the equation in the paper,
the RRM predator fitness function requires that there is two prey and
throws an exception if there is not. The RRM prey fitness function
should work with any number of prey or predators