PredPrey: Created another pred fitness function, fixed RRM functions, cleaned up batch files
Created a version of the predator fitness function from the RRM paper
which is generalized. It has the same idea in that it emphasizes first
the percentage of prey that are caught, then factors in the distance of
the prey from the predators to minimize that distance. My function is
called PredatorCatchCloseObjective. Also fixed the bugs from both the
prey and predator RRM fitness functions. Also cleaned up the batch files
so that they always run the best static controller with the same
constant variables with each evolved fitness function agent in order to
compare them all cleanly.