comments added for bug testing JUnit test
this is a mess guys, bear with me here. The JUnit test for
TetrisAfterStateGenerator stopped working after we made our updates to
the code (uses BFS now). We are commenting this to keep track of these
bug testing print statements, but they will likely be removed after the
JUnit test is fixed.