Fixes to ImageMatch resolution
Removed large French Flag again ... we really don't need it. That's the
whole point of the CPPNs having infinite resolution. I changed the main
method of the ImageMatchTask to demonstrate this. It takes a single CPPN
and uses it to draw the same image at 3 different sizes. I also changed
how the image size and override parameters work. If the size is not
being overridden, then watching the result from ImageMatch will be the
same size as the original image. However, if the size is overridden,
then the new width and height are used by the CPPN to draw its image. It
is important to note that the CPPN was only trained to match the
original smaller image. The fact that you can rescale the image after
the face is just a cool feature of CPPNs.