Batch file organization: Ms-Pac Man
I moved all batch files associated with Ms. Pac-Man experiments to a new
subdir in the new batch directory. Every file had to have the commands
cd ..
cd ..
added to the start to make sure code was launched from the correct
working directory, but this was all that was required. This change
should clean up our main directory a bit and keep things more organized.