Many HyperNEAT in Pred/Prey changes
I cleaned up a lot of the new HyperNEAT code for the pred/prey domain,
and also added some bits of functionality that we had forgotten to try
so far. However, there is still an error. Start addressing this by
running the newly added
PredPreyTorus-CatchCloseQuickPredVsStaticPreyTeam-HyperNEAT.bat and
troubleshooting from there.