Added some comments and a JUnit test case to a couple general classes, generalized predPrey sensors
Commented CooperativeCoevolutionMuLambda so that the class has a comment
explaining in a general sense every method and class. Replaced the main
method in the Interval class with unit tests. Generalized the sensors
for the predPrey agents and made the code a lot more compact and elegant
(and commented it). Added a new sensor package so that all future sensor
blocks can be added in much more simply, so sensors by proximity can be
added next.