Changes to allow multiple RLGlue instances
We changed the Runtime.exec() to ProcessBuilder.(command).start() and
initialized RLGlue with a port number before calling it in the method
prep(). All in all, this means we now have a parameter "rlGluePort" that
we can change in the batch command line files so we can run multiple
RLGlue instances at once.