Don't track combining crossover information
All tracking of combining crossover information is now optional, and
disabled by default (can be turned on with command line parameters).
This change only prevents the writing of a single file, but there was
also a lot of overhead involved in that file because using combining
crossover also meant that the archetype had to be managed in special
ways. I believed that I didn't remove anything that would break
anything, but we'll wait and see.