HyperNEAT overrides
HyperNEAT seems to work, but is very slow. It needs some special
handling.

For example, using monitorInputs is inappropriate with HyperNEAT because
the sensor and output labels of the NetworkTask interface only
correspond to "normal" network inputs and outputs.

Also changed the HyperNEAT-PredPrey batch file to use smaller number in
hopes of watching how evolution progresses.