Added the PreyMinimizeCaughtObjective and parameters to run the other two new ones
Added PreyMinimizeCaughtObjective. Now there are mirror image fitness
functions between the predators and the prey of the following based
fitness scores: total game time, number of prey caught, speed of each
prey's death, distance between the predators and the prey, and the
fitness functions implemented exactly (generalized as much as possible
to allow for variable numbers of prey and predators) from the RRM paper.
Other than these 10 fitness functions, there's just one more for the
predator called the PredatorCatchCloseObjective which is a remade
version of the RRM paper predator fitness function because I was unable
to generalize their predator fitness function (in the paper it was hard
coded for 2 prey). This remaking process was not required for the prey
RRM fitness function because it was able to be generalized.