Simplified/rearranged some VizDoom code
The getInputs method is simplified to only loop through the desired row
(looping through y values was not needed).

The row number is now calculated once in the constructor and saved
rather than recalculated every time inputs are calculated.

Note the comment in VizDoomTask: all reference to the getRow() method
should eventually be removed from this class.