Changes in how HyperNEAT is sped up
The main change is linkGeneIndex, but there are also formatting changes.
Also, I made the cached information static ... it isn't any use if we
are recalculating it for each genotype. The whole point is that
different genotypes are making phenotypes with the exact same
structure..

This all needs to be extensively tested still.