UT2004 Doesn't work any more?
I cleaned up some UT2004 and added a batch file because I wanted to test
it out, but the old does doesn't work any more. It seems to be able to
launch a server just fine, and can even add native bots to it, but has
trouble connecting GameBots agents.  There is a problem parsing a Yylex
message early on: State MSG_INV unprocessed. The message is
"{Reachablefalse}". There should probably be an = in between, but why is
this error popping up now? None of the Pogamut or GameBots stuff should
have changed. Try this on a different system and see what happens.