Created better distance fitness function to fix incorrect logic of previous distance fitness for Predators
Created distance fitness function to each individual prey, called
PredatorMinimizeDistanceFromIndividualPreyObjective, instead of the prey
as a team so that the predators do not learn to minimize distance to
prey all at once, essentially avoiding actually catching prey because
they are moving towards the closest possible middle distance between the
prey.
Created corresponding batch file, parameter, and task updates to include
the new function.
Began allowing for "otherStats" to be kept in PredPrey which don't
effect evolution