Some changes to Picbreeder
I rearranged the code in this task a bit to nudge it in the right
direction. However, even the basic main method still does not draw the
image buttons. I'm actually not sure why this is, but this obviously
needs to be resolved before things like score calculation, etc can be
figured out.