Attempted to finish implementation of CooperativePredPrey, still needs work
Attempted to finish implementation of CooperativePredPrey, still needs
work. Currently when the coevolution task is run, there is a null
pointer exception at CooperativeCoevolutionMuLambda line 76, which
refers to the size of genotypeExamples in MMNEAT. I assume that it is
necessary to implement a utility method for the cooperative coevolution
tasks which sets up the coevolving task, just like cooperative
coevolving mspacman tasks in the MMNEAT class, but I am unsure how to do
this.