Picbreeder almost there
Got it more or less working. Wasn't completely sure where to put the EA
code, made a note of it in the class. I *think* it's working though. The
only thing is that the clicks only seem to register on every other run,
but I'm not sure if that's due to my sketchy testing method or the code
itself. I think I'm going to hold off on fiddling with these specific
issues until I can talk some more with Schrum