PredPrey: Created four very specific fitness functions (2 pred and 2 prey), fixed a bug, added parameters.
PredPrey: Created four very specific fitness functions (2 pred and 2
prey), fixed a bug, added parameters. The four fitness functions are
meant for cooperative coevolution, along with command line parameters
for each of the functions, but they could also be used for any predPrey
task, and may be desirable to try in the TorusPredPreyTask vs static
controllers because of their simplicity and multi-objective properties.
They are not yet being used in the code. Also, I fixed a bug with the
PredatorMinimizeDistanceFromIndividualPreyObjective which would cause
the minScore to be incorrect.