Many VizDoom Fixes
This fixes several VizDoom issues and cleans up the code a bit. The
Defend the Center task now works with both the visual and non-visual
approach. However, I simplified the inputs to be the same as used in the
basic shoot task. In fact, these are now default for all tasks and
accessible via that class from static methods (less duplicated code).
This closes #138 .

The reason for the watch being forced seems to be the loading of the CFG
file, and the order in which it was done. I switched around commands so
that the CFG file for the scenario will be loaded first (each sub class
of VizDoomTask should override this), and then all other setting
commands are run, which overrides the CFG files.