PredPrey: new batch file, better otherScore ordering, ran experiment
Created the batch file for multiObjective CCQ but with the distance
function encouraging being close to each individual prey instead of the
prey as a team. Moved the otherScore for the predatorCatchObjective to
the top so that it will come up first (after the actual fitness
functions being used) because it is the most telling. Then, I ran the
experiment on the server for each of the three predator individual
distance functions and for the multiCCQ with individual distance
function (the batch file I just made) for a total of 4 experiments.