PredPrey: attempted to implement the saving of the best agent in each objective in each population
Attempted to implement the saving of the best agent in each objective in
each population in torusPredPrey cooperative coevolution. I was able to
implement all of the code (and I think it will work correctly) except
for the small but essential part of getting the score of each agent, and
because I was unable to do this the rest of the code that I just added
needed to be commented out as well for now.