changed parameters for picbreeder task to simplify them
took out the numRows and numColumns and just takes mu as a parameter.
The way I have it currently written though, mu does get changed so all
the buttons in the frame are filled. Also couldn't figure out the
scroller issue, but not sure we should really worry about that. The
original picbreeder hardcodes the size of the population and the way I
have it written right now the population is coded to be anything that is
a multiple of 4 and can fit on the screen.