got the draw networks button up and running!!
if network button clicked, draws network onto button. then, if it's hit
again, it reverts to the images. Also, if you hit the evolve button with
networks on, it will show the next generation as networks. Only issues
are that images aren't quite to scale when network button clicked, so
you cant see the inputs or outputs. That's due to the resettable picture
size so not sure how to fix this without removing that feature... Also,
there's this odd bug where if you click on a few pictures, then click
the network button to see the networks, but the unclick it to see the
pictures again, then immediately try and hit the evolve button, it
doesn't do anything. However, if you click on another image button, that
seems to wake up the program and then it evolves just fine. Strange bug
though... Also have it so it saves the network if the network is
currently being displayed. Need to decide if that's what we want though