Made standard lineage browser work with Picbreeder
Fixed a lot of my code in the poorly named Offspring class (this is
where all the lineage browsing happens). For this to work properly, I
had to fix a some of the lineage logging (wrong parent ids were used)
and create special cases for Picbreeder. I also changed
SelectiveBreeding to output fitness logs again. They are not very
informative, but the lineage browser wants them and it's easier just to
keep them.