Updated batch files. Fixed some PredPrey errors, cleaned up code.
Made the cooperative predPrey batch files have better names. Made a
batch file for the postBestObjectiveWatch for cooperative predPrey and
another one for the Eval for cooperative predPrey. Fixed a few bugs in
the ObjectiveBestTeamsExperiment class, although there is now a new
error on line 63 when trying to run postBestWatch, saying I can't cast
the array to a Genotype[] array. Also, there is an error with the new
post best eval batch file on line 48 of ObjectiveBestTeamsExperiment.
Added a parameter for the  cooperative watch. Fixed several errors in
the CooperativeTask, which now reveals another new error on line 190 of
TorusPredPreyTask when running at the end of the first generation with
monitorInputs and watch turned on. This error is because it uses the
lonerTask "evolved" array which is null for cooperativeTask, but I am
unsure as to how to avoid this error. Cleaned up
CooperativeTorusPredPreyTask a bit