Improved lineage option of picbreeder
Reduced size of lineage images and changed the way they display. Also
maintain images in lineage in a set so that duplicates are not printed
(which was overwhelming the screen).

However, there still needs to be a way to close these image windows
without terminating the program.

This commit contributes to #151