Added Ability to Smudge in VizDoom
I've added a new parameter and new methods and checks to VizDoom to
allow BasicShooterTask to smudge inputs on  given section. Right now, if
"doomInputPixelSmudge" is greater than 1, it will affect the inputs
width and height and smudge together groups of pixels. It greatly
reduces the number of inputs needed for a section,