Created some batch files for implementing post best watch/eval for a designated coevolution team.
Created a lot of batch code which can concatenate and go through a
variable number of arguments. Provided this code in an example batch
file called "batchFileGuineaPig.bat". Began to translate this to
implementing the batch files for postBestWatch and postBestEval for a
given cooperative team. postBestObjectiveWatchTeam.bat most likely
works and is fully implemented.