Generalized launchMultiple
Generalized the launch multiple batch file so that the user specifies
the starting and ending run numbers. Also allowed the batch file to be
able to handle two different sets of experiments in one run. Now I will
create a batch file that will do all of the same things as this one,
except that it will run the experiments asynchronously rather than
simultaneously