Generalized the objectives list for the predPrey to be able to support multiple populations
Generalized the objectives and other scores lists in the predPrey world
so that it is a list of lists, so that it can essentially contain
various lists of objectives for each different population for
coevolution. Currently, the code is still the same by just using only
the 0 index of the double arrayList and putting everything as normal in
that index as an arrayList, but now all of this code can be used by
coevolution to make different sets of fitnesses for each population. I
also compiled the adding of all of the objectives into a single method
in the parent torusPredPreyTask that is simply called by the predator
and the prey tasks and can be used later by cooperative coevolution. I
also added a ToDo to make the double arrayList counterpart for
objectives in the main MMNEAT class.