Altered rlGlue batch files so that the port number will be different for each experiment run
Changed the RLGlue batch files so that each experiment that is run will
be run on the rlGluePort 4096+the run number. Note: if different
experiment types of the rlGlue are run simultaneously, this won't work
currently (still) because the port numbers will conflict. The solution
might be to start each port number at ten more in each batch file, so
that each batch file is allowed a gap of 10 ports (for instance, the
first could be 4096+run number, the next would be 5006+run number, and
so on) given that we have enough port numbers to support this. Also
updated the comments in the launchMultiple batch files so that they
represent a batch file that actually exists.