Fixed problems causing postProcess to not work with coevolution.
Fixed problems causing postProcess to not work with coevolution.
Specifically, it now works with cooperative coevolution with the
torusPredPrey. Some work arounds had to be done and may need to be
changed later, especially once various sets of fitness functions for
each population are implemented. These work arounds are marked with TODO
comments in the code.