Fixed an error causing the cooperative coevolution not to run.
Fixed an error causing the cooperative coevolution not to run, found
when attempting to generalize the code to add objectives for each
population rather than just the first. This was because the otherScores
were only being given to the first population so when the code attempted
to read in the otherScores for the second population, there was an index
out of bounds error. Fixed by making only the first population in
otherStatsPerPopulation have the number of otherScores and having all of
the other populations have a size of zero.