Generalized the fitness functions ArrayList to support coevolution (now its a double ArrayList)
Generalized the fitness functions ArrayList to be a double ArrayList so
that it can support coevolution with different sets of fitness functions
for each population. In order to support non-coevolution, the original
single list of fitness functions, which is copied for each population,
is just placed into the first index of the double arrayList in the case
that coevolution is not happening.