PredPrey; made new fitness function for prey, created new experiment for cooperative coevolution with different fitnesses per population
Created a new fitness function for the prey which takes an index for the
individual prey to score for and gives it a score based on maximizing
its distance from the closest predator to that prey. Added this new
fitness to the prey otherScores. Created a new parameter for the
predator and a new parameter for the prey for a cooperative coevolution
setup giving the predator/prey multi CCQ to each agent individually and
giving the predator/prey only the distance function(s) that correspond
to that agent itself, not the whole team (so the predators and the prey
will have different sets of fitness functions, although still the same
number of objectives). Created batch files to run each of these new
experiments, and ran these experiments on the server along with the
other, original cooperative ones that have the same fitness functions.
NOTE: (very important): still need to check that various sets of fitness
functions will work with varying number of objectives, I suspect they
probably won't.