Added Super Mario!
If you go to ch.idsia.scenarios and run the Play class, you should be
able to play Super Mario. However, the next step is to actually make a
task to evolve in Mario. The code actually seems to already have a
built-in example of how to perform evolution, so we can look at that
code when creating a Mario task to work with our system.

This closes #127, but new related issues will soon be made.