PredPrey: new herding fitness for preds. New coop experiment. Fixed errors with having various fitnesses per population
Created new fitness function for the predators for herding the prey
(minimizing distance between the prey). Created a cooperative coevolving
experiment with the evolved predators where the first predator is given
this herding fitness and all other predators are given multi CCQ with
the distance for that specific agent. Fixed errors across many classes
preventing various numbers of fitnesses per populations from working
(log, MMNEAT register fitness, etc). Now the
ObjectiveBestTeamsExperiment and the postProcessing needs to be
generalized to work for various numbers of fitnesses per population
(coevolution in general) because right now they just assume that the
first population's objectives are used for every population.