Implemented competitive coevolution
Implemented competitive coevolution. It extends the cooperative
coevolution task in the torusPredPrey world, so in the future I will
correct all of the cooperative references that are being used now by
both competitive and cooperative coevolution. These false references are
in both the torusPredPrey domain and also scattered throughout the
overall coevolving domain (nsga thing). #155