Updated other scores for competitive coevolution in predPrey
Made otherScores hold both prey and predator fitnesses (all possible) in
competitive coevolution in the first population (regardless of whether
that population is a predator or a prey). Made it where if it is
non-competitive coevolution, it will still only add all fitnesses to the
other scores for the appropriate agent type.