PredPrey new experiment/parameters/batch
Created experiment for cooperative coevolution based off of the best
result from the last experiment, which was coOp multi CCQ with indiv
distance to each prey for each population-specific predator (each
population only has the distance function for that population), except
without the "quick" fitness part of it. This was done in order to more
closely follow the RRM paper that this is all based off of (distance and
caught fitness only) but with multi-objective rather than a single more
complex fitness (and with cooperative coevolution). I am also hoping
that this will cause more complex behavior (such as herding) instead of
rushing to catch the first prey then the second prey as quick as
possible.