MLP and SubstrateMLP
I fixed the NetworkUtilTest, but SubstrateMLP still has problems. I did
find some genuine errors, but they were small. The main issue was that
my TWEANN activates input neurons before they transmit, which is
actually NOT the standard neural network model, but should be harmless
to do. However, the SubstrateMLP was wrong because it didn't take the
quirk of my TWEANN into account. The other changes I made to Substrate
MLP were just for tidiness.

However, SubstrateMLP and TWEANNs generated by CPPNs are still not
compatible, and I know why: the structures are slightly different
because of reflections across the diagonals of the substrates. I think
the error might actually be in the HyperNEATCPPNGenotype in how it
matches up innovation numbers with nodes. They seem to count down across
columns rather than across rows. However, a simple attempt at fixing
this did not immediately solve the problem, so I reverted changes to
that class for now.

Need to look closely at both SubstrateMLP and HyperNEATCPPNGenotype. For
now, I've left some commented debugging code in certain classes.

This relates to issue #188