Fixed issues with HyperNEATCPPNGenotype and SubstrateMLP
The HyperNEATCPPNGenotype was actually initializing the innovation
numbers wrong so that they were organized in column-major order rather
than row-major order. This means that inputs and connections would have
been set up in a weird way that didn't correspond to the coordinate
frame.

Additionally, the SubstrateMLP had similar but different problems. It's
links were being established correctly, but the inputs were being loaded
incorrectly. Additionally, the outputs were being extracted incorrectly.
This once again has to do with the distinction between row- and
column-major ordering.

At this point, but the TWEANNs generated by HyperNEATCPPNGenotype and
the SubstrateMLP are identical in their behavior ... at least, the
testProcess test of SubstrateMLPTest passes. The valuesTest still fails,
but I'll leave that to Lauren.

There are also pieces of commented print statements all over the place
that should be cleaned up, but that can be left for later.

This commit applies to #188