More steps towards Mario HyperNEAT
I had to hard code that there will always be two input substrates, one
for the world and one for the enemies. If you want this more generalized
I can go back and fix this, but I don't think we'd use anything else or
less? Also, I've hardcoded the Mario output buttons to 5, and I'm fairly
certain that is doesn't break anything, but I'm not sure if you want
that to be hardcoded or not. Using the keys gives us 16 instead of 5.