PredPrey: Created all official experiments in the code and batch files
Code made ready to run all official experiments for torusPredPrey
IndividualVsTeam Selection by adding the objectives setup in the
torusPredPreyTask, Parameters for each setup (along with explanations),
and batch files. Added "cooperativeBalancedIndividualAndTeamSelection"
which will be a selected balance of individual and team selection with
individual distance selection and prey caught team selection (used to be
predatorCoOpCC). There is a total of ten experiments, although one (the
cooperativeIndividualAndTeamSelection setup) is most likely unnecessary
and doesn't need to be run.