NetGlue -> LocalGlue
This whole time I've been using the NetGlue codec for RL-Glue, which
sets up three separate threads which communicate using sockets. However,
it turns out this is completely unnecessary and wastefully slow. There
is a LocalGlue codec that just runs everything in a single thread
without needing any socket communication. It also does not require the
RLGlue.exe program, which means that we should be able to run this code
on non-Windows systems now.