Actually removed rl_glue.exe!
Since we use the LocalGlue codec now, we don't need this executable, or
any of the code that uses it. This simplifies the code, and removed the
need to include a compiled executable in the repository. Still need to
remove the rlGluePort parameter from the code and batch files though.