VisualNode fixes
Node sorting order is fixed, and I also made it so that the color is
different from displaying the maximal node. Also, the maximal node now
shows all ties, which revealed what might be a huge problem ... lots of
ties! Or is this just because I'm looking at networks from the first
generation?