Almost got weight visualization working
Got everything(I think) The only thing is there’s a method in the
hyperNEATCPPNGenotype that I wasn’t sure how to write or why it was
necessary. Also, Im not sure how to set the color and the grid size to
draw the node in the drawingPanel. Ive left todos in all of the classes
that I still have questions on.