Implemented getSubstrateGenotype()
This method gets the TWEANNGenotype representation of the substrate
network generated by a CPPN. This code was in the getPhenotype() method,
but was moved to this new method to allow the TWEANNGenotype rather than
the TWEANN itself, which is what the getPhenotype() method returns. In
fact, the getPhenotype() metho simply calls the new
getSubstrateGenotype() method, and then returns the phenotype from the
resulting genotype.

This change should given easy access to TWEANNGenotype methods, allowing
easier addressing of issue #246