Comments regarding weight viz errors
The weight visualization still has errors, but I've put in some code
that will make it easier to troubleshoot. The current problem is an
array index out of bounds error ... this might be as simple as an
off-by-one error, or it could be something bigger. Too big to delve into
it now.

Look at my comments, and take advantage of the exception catching
mechanism that only prints out debugging output when an error occurs.

#246