HyperNEAT Weight Viz Complete!
I think I've fixed the weight visualization for HyperNEAT! We need to do
tests with different substrates and different domains, number of
processing layers, and extra links, etc, but it seems to work.

Your code was pretty close. Some s1 vs s2 variables had to be switched
to rotate things some more, and > had to be changed to >= in two places,
which seems to take care of it.

#246