Give "frozen" protected access
I made the frozen boolean field of the Gene class protected, which
necessitated the use of a public method in order to access it. Besides
being good coding style, this will also make it easy to comment out the
frozen field to see if the removal of one small field from over a
million Gene instances can change the memory impact of large networks.
#254