Smaller networks with HyperNEAT seeded tasks
When a HyperNEAT task is used to seed the initial population, a boolean
is set that makes TWEANNGenotypes use a different type of NodeGene and
LinkGene. Specifically, several optional fields are not included, which
allows the size (in memory) of the networks to be smaller.

For example, in trial runs in Tetris, the size of network xml save files
drops from 38,860KB to  26,069KB. More could be done if the link weights
were switched from doubles to floats, but I fear that this could
actually affect evolution performance.

Despite this change, I might still have memory problems. A lot of my
profiling tests indicated that memory is leaking TWEANNGenotype
instances ... there are more Live instances than there should be per
generation.

Testing must continue, but having the option for smaller networks helps.

#254