static Gene classes
I made the NodeGene and LinkGene classes static so that they are not
implicitly tied to instances of TWEANNGenotype. I think this allows
certain TWEANNGenotypes to be garbage collected that could not be
collected before, though I'm not entirely certain. This version of the
code seems functionally identical, so I'm going to keep the changes.
#254