Moved culling code into own method
Profiling has led me to crossover as the cause of the memory leak of
TWEANNGenotypes. I plan to clean up NSGA2 and/or the TWEANNCrossover
classes in order to fix this, but wanted to organize the existing code a
bit first. This is why the culling code has been moved out into its own
method. #254