Memory leak fixed in TWEANN Crossover
Instead of creating a new genotype inside of the crossover method, the
toReturn parameter is both modified and returned. This should not cause
any problems because the genotype sent in as a parameter here is a copy
of a genotype and not an actual member of a parent population.

I have verified with the profiler that this new approach does not cause
a memory leak, but I need to verify via other means that it does not
cause other problems. Check JUnit tests, do some short evolution runs in
several domains, etc.

Also, this test revealed that crossover is extremely slow. From looking
at the code, I think there are some ways to speed it up a little bit.

#254