Added methods to HyperNEATTask
The HyperNEATTask now has methods to get the number of CPPN inputs and
to filter the "default" set of inputs containing x1,y1,x2,y2,bias. This
will allow for 1D substrates ( #266 ).

Currently, every HyperNEAT task uses the default definitions of these
methods, which correspond to 2D substrates. The next step is to modify
the HyperNEAT Tetris task to actually change the number of inputs and
filter the default input set, when appropriate.