Fixed HyperNEAT Bias calculations
Intead of having a bias output for every pairing of connected
substrates, there is simply one bias output per non-input substrate.
Also, the correct outputs are actually used to assign bias values to the
substrates now. This resolves #265 but leads to a new issue, which is
properly testing these changes: #283