Static global variables made private, getters added in affected methods
Changed all static member variables in Substrate.java to be private
instead of public. This caused errors in all of the methods that used
these public variables, but these were fixed when I changed the variable
instances to the getter method calls. All methods are functioning as
they were originally. close #281