#261 Option to Turn Off Each Picbreeder Input
Slight edits to names of the Effect CheckBoxes to ensure correct
labeling. Effect CheckBoxes were restored, and functionality was added.
Method to reset the existing images was added, as well as a method to
handle Effect CheckBox activation or de-activation. The previously
existing setCheckBox method was renamed to setActivationFunctionCheckBox
to avoid future confusion.