JLayer library for MP3 playback
I added the JLayer library as a jar file to support fairly easy MP3
playback. There is an example in SoundUtil, but it needs to be broken
down into more usable methods.

JLayer supposedly has both an MP3 encoder and decoder, which should
allow us to convert back and forth between mp3 and sequences of numbers.
Look at the classes in the jl1.0.1.jar library, and also search the
internet for "JLayer encode" or "JLayer decode" to find examples of how
to do this.

#271