Can see mountain and goal
The Mountain and Goal now show up in the visualization, but not the car.
The problem with the mountain and goal were  that a scaling
transformation was supposed to be used, but was causing problems. All
coordinates were given within a unit square, but when trying to draw
that in the DrawingPanel, nothing showed up. I had to scale everything
up myself.

I think I can add the car by stealing some render code from another viz
component. Will hopefully commit that soon. #118